separate base types
